feat

+++ Transactions/Resources/Create/CreatePlanResource.cs
@@ -4,6 +4,6 @@ public class CreatePlanResource
 {
     public string? Name { get; set; }
     public string? Description { get; set; }
-    public long Fee { get; set; }
+    public double Fee { get; set; }
     public int SupplierId { get; set; }
 }
